requirements.txt
streamlit
google-generativeai
PyPDF2
python-pptx
pandas